+++ src/mimic/src/orm/mod.rs
@@ -12,7 +12,7 @@ pub mod prelude {
         ic::structures::storable::Bound,
         orm::{
             OrmError,
-            base::types::Ulid,
+            base::types::{Ulid, UlidSet},
             traits::{
                 EntityDyn, EntityFixture, EntityId as _, Filterable, Inner as _, NumCast,
                 Orderable, Path, Selector as _, SortKey as _, Storable, Validate as _,